Support to import types
