+++ package.json
@@ -3,7 +3,7 @@
   "private": true,
   "version": "0.1.0",
   "description": "Firebase Authentication for Google Apps Scripts",
-  "main": "index.js",
+  "main": "src/FirebaseAuthentication.ts",
   "repository": "https://github.com/tricknotes/FirebaseAuthenticationGoogleAppsScript",
   "author": "Ryunosuke Sato <tricknotes.rs@gmail.com>",
   "license": "MIT",
